fix(backoffice-v2): fixed default open state on nav items and active state
