+++ apps/backoffice-v2/src/pages/CaseManagement/CaseManagement.page.tsx
@@ -1,3 +1,8 @@
 import { Outlet } from 'react-router-dom';
+import { useSelectEntityFilterOnMount } from '@/domains/entities/hooks/useSelectEntityFilterOnMount/useSelectEntityFilterOnMount';
 
-export const CaseManagement = () => <Outlet />;
+export const CaseManagement = () => {
+  useSelectEntityFilterOnMount();
+
+  return <Outlet />;
+};
